Refactor: #56 - model의 필드 타입 변경
Stadium id 는 Long 타입으로 되어 있는데 fk인 team_id는 Integer로 되어 있어서 수정함

+++ src/main/java/team/snowball/baseball/model/team/Team.java
@@ -17,7 +17,7 @@ import java.sql.Timestamp;
 public class Team {
 
     private Long id;
-    private Integer stadiumId;
+    private Long stadiumId;
     private String name;
     private Timestamp createdAt;
